feat: features/siginup signup logic 90%
